+++ Omega/res/raw/changelog.md
@@ -1,4 +1,6 @@
 ### v0.5.0 (Build 40)
+* **Add: Desktop grid preference**
+* **Add: Desktop Hide Labels preference**
 
 ### v0.4.0 (Build 32)
 * **Add: Hide apps preference**