+++ Omega/res/raw/changelog.md
@@ -2,6 +2,7 @@
 * **Add: Desktop grid preference**
 * **Add: Desktop Hide Labels preference**
 * **Add: Desktop Multiline labels preference**
+* **Add: Dock icons count preference**
 
 ### v0.4.0 (Build 32)
 * **Add: Hide apps preference**