+++ Omega/res/raw/changelog.md
@@ -4,6 +4,7 @@
 * **Add: Desktop Multiline labels preference**
 * **Add: Dock icons count preference**
 * **Add: Swipe down gesture**
+* **Add: Predictions preference**
 
 ### v0.4.0 (Build 32)
 * **Add: Hide apps preference**