+++ Omega/res/raw/changelog.md
@@ -5,6 +5,7 @@
 * **Add: Dock icons count preference**
 * **Add: Swipe down gesture**
 * **Add: Predictions preference**
+* **Add: Search bar radius preference**
 
 ### v0.4.0 (Build 32)
 * **Add: Hide apps preference**